+++ src/stores/AppStore.ts
@@ -1108,7 +1108,7 @@ export class AppStore {
         try {
             const ack = await this.backendService.requestFitting(message);
             if (ack.success) {
-                this.getFrame(message.fileId).updateFittingResults(ack.resultValues, ack.resultErrors, ack.log);
+                this.imageFittingStore.setResultString(ack.resultValues, ack.resultErrors, ack.log);
             }
             if (ack.message) {
                 AppToaster.show(WarningToast(`Image fitting: ${ack.message}.`));
